+++ src/pages/onboardingv2.js
@@ -6,6 +6,7 @@ import { CippWizardOptionsList } from "../components/CippWizard/CippWizardOption
 import { CippSAMDeploy } from "../components/CippWizard/CippSAMDeploy.jsx";
 import { CippTenantModeDeploy } from "../components/CippWizard/CippTenantModeDeploy.jsx";
 import { CippBaselinesStep } from "../components/CippWizard/CippBaselinesStep.jsx";
+import { CippNotificationsStep } from "../components/CippWizard/CippNotificationsStep.jsx";
 import { BuildingOfficeIcon, CloudIcon, CpuChipIcon } from "@heroicons/react/24/outline";
 
 const Page = () => {
@@ -73,24 +74,18 @@ const Page = () => {
     },
     {
       title: "Step 5",
-      description: "Integrations",
-      component: CippDeploymentStep,
-      //give the choice to configure integrations.
+      description: "Notifications",
+      component: CippNotificationsStep,
+      //Show the notification menu (cipp/settings/notifications) without the submit/save button, but with a test button.
     },
     {
       title: "Step 6",
-      description: "Notifications",
-      component: CippDeploymentStep,
-      //explain notifications, test if email is setup,etc.
-    },
-    {
-      title: "Step 7",
       description: "Alerts",
       component: CippDeploymentStep,
       //show template alerts, allow user to configure them.
     },
     {
-      title: "Step 8",
+      title: "Step 7",
       description: "Confirmation",
       component: CippWizardConfirmation,
       //confirm and finish button, perform tasks, launch checks etc.
